chore: extract styles of modal steps to their own css module file
