+++ src/client/v2/src/components/AddAccountModal/steps/SelectAccountTypeStep.module.css
@@ -1,0 +1,11 @@
+.accountTypeCard {
+  flex: 1;
+  padding: var(--mantine-spacing-xl);
+  border-radius: var(--mantine-radius-md);
+  cursor: pointer;
+}
+
+.accountTypeCard[data-checked] {
+  background-color: light-dark(var(--mantine-color-blue-0), var(--mantine-color-dark-6));
+  border-color: var(--mantine-color-blue-filled);
+}
